# Rust performance improvements