--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,5 @@ target
 .cache
 compile_commands.json
 .luacheckcache
+benchmarks
+flamegraph*