--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ compile_commands.json
 .luacheckcache
 benchmarks
 flamegraph*
+perf.data*