--- a/examples/filename_search.rs
+++ b/examples/filename_search.rs
@@ -1,5 +1,7 @@
 use ivyrs::inner_files;
 
 pub fn main() {
-    inner_files("file.go".to_owned(), "/tmp/ivy-trees/kubernetes".to_owned());
+    let res = inner_files("file.go".to_owned(), "/tmp/ivy-trees/kubernetes".to_owned());
+
+    println!("{}", res);
 }