--- a/rust/lib.rs
+++ b/rust/lib.rs
@@ -1,30 +1,35 @@
-mod matcher;
 mod finder;
+mod matcher;
 mod sorter;
 mod thread_pool;
 
-use std::sync::Mutex;
 use std::collections::HashMap;
-use std::os::raw::{c_int, c_char};
-use std::ffi::CString;
 use std::ffi::CStr;
+use std::ffi::CString;
+use std::os::raw::{c_char, c_int};
+use std::sync::Mutex;
 
 #[macro_use]
 extern crate lazy_static;
 
 lazy_static! {
-    static ref GLOBAL_FILE_CACHE: Mutex<HashMap<String, Vec<String>>> = return Mutex::new(HashMap::new()) ;    
+    static ref GLOBAL_FILE_CACHE: Mutex<HashMap<String, Vec<String>>> = Mutex::new(HashMap::new());
 }
 
 fn to_string(input: *const c_char) -> String {
-    return unsafe { CStr::from_ptr(input) }.to_str().unwrap().to_string();
+    unsafe { CStr::from_ptr(input) }
+        .to_str()
+        .unwrap()
+        .to_string()
 }
 
 fn get_files(directory: &String) -> Vec<String> {
     let mut cache = GLOBAL_FILE_CACHE.lock().unwrap();
     if !cache.contains_key(directory) {
-        let finder_options = finder::Options{ directory: directory.clone() };
-        cache.insert( directory.clone(), finder::find_files(finder_options));
+        let finder_options = finder::Options {
+            directory: directory.clone(),
+        };
+        cache.insert(directory.clone(), finder::find_files(finder_options));
     }
 
     return cache.get(directory).unwrap().to_vec();
@@ -41,8 +46,9 @@ pub extern "C" fn ivy_match(c_pattern: *const c_char, c_text: *const c_char) -> 
     let pattern = to_string(c_pattern);
     let text = to_string(c_text);
 
-    let m = matcher::Matcher::new( pattern );
-    return m.score(text) as i32;
+    let m = matcher::Matcher::new(pattern);
+
+    m.score(text) as i32
 }
 
 #[no_mangle]
@@ -52,7 +58,7 @@ pub extern "C" fn ivy_files(c_pattern: *const c_char, c_base_dir: *const c_char)
 
     // Bail out early if the pattern is empty its never going to find anything
     if pattern.is_empty() {
-        return CString::new("").unwrap().into_raw()
+        return CString::new("").unwrap().into_raw();
     }
 
     let files = get_files(&directory);
@@ -66,6 +72,5 @@ pub extern "C" fn ivy_files(c_pattern: *const c_char, c_base_dir: *const c_char)
         output.push('\n');
     }
 
-    return CString::new(output).unwrap().into_raw()
+    CString::new(output).unwrap().into_raw()
 }
-