--- a/rust/matcher.rs
+++ b/rust/matcher.rs
@@ -1,5 +1,5 @@
-use fuzzy_matcher::FuzzyMatcher;
 use fuzzy_matcher::skim::SkimMatcherV2;
+use fuzzy_matcher::FuzzyMatcher;
 
 pub struct Matcher {
     /// The search pattern that we want to match against some text
@@ -9,17 +9,16 @@ pub struct Matcher {
 
 impl Matcher {
     pub fn new(pattern: String) -> Self {
-        return Self {
+        Self {
             pattern,
             matcher: SkimMatcherV2::default(),
         }
     }
 
-    pub fn score(self: &Self, text: String) -> i64 {
-        if let Some((score, _indices)) =  self.matcher.fuzzy_indices(&text, &self.pattern) {
-            return score;
-        }
-
-        return 0;
+    pub fn score(&self, text: String) -> i64 {
+        self.matcher
+            .fuzzy_indices(&text, &self.pattern)
+            .map(|(score, _indices)| score)
+            .unwrap_or_default()
     }
 }