--- a/rust/sorter.rs
+++ b/rust/sorter.rs
@@ -1,8 +1,8 @@
 use super::matcher;
 use super::thread_pool;
 
+use std::sync::mpsc;
 use std::sync::Arc;
-use std::sync::Mutex;
 
 pub struct Match {
     pub score: i64,
@@ -23,30 +23,37 @@ impl Options {
     }
 }
 
-pub fn sort_strings(options: Options, strings: Vec<String>) -> Arc<Mutex<Vec<Match>>> {
-    let matches: Arc<Mutex<Vec<Match>>> = Arc::new(Mutex::new(Vec::new()));
-    let matcher = Arc::new(Mutex::new(matcher::Matcher::new(options.pattern)));
+pub fn sort_strings(options: Options, strings: Vec<String>) -> Vec<Match> {
+    let mut matches = Vec::new();
+    let matcher = Arc::new(matcher::Matcher::new(options.pattern));
 
     let pool = thread_pool::ThreadPool::new(std::thread::available_parallelism().unwrap().get());
 
+    let (tx, rx) = mpsc::channel::<Match>();
+
     for string in strings {
         let thread_matcher = Arc::clone(&matcher);
-        let thread_matches = Arc::clone(&matches);
+        let thread_transmitter = tx.clone();
         pool.execute(move || {
-            let score = thread_matcher.lock().unwrap().score(string.to_string());
+            let score = thread_matcher.score(string.to_string());
             if score > 25 {
-                let mut tmp = thread_matches.lock().unwrap();
-                let content = string.clone();
-                tmp.push(Match { score, content });
+                thread_transmitter
+                    .send(Match {
+                        score,
+                        content: string,
+                    })
+                    .expect("Failed to push data to channel");
             }
         })
     }
 
     drop(pool);
+    drop(tx);
 
-    matches
-        .lock()
-        .unwrap()
-        .sort_by(|a, b| a.score.cmp(&b.score));
+    while let Ok(result) = rx.recv() {
+        matches.push(result)
+    }
+
+    matches.sort_by(|a, b| a.score.cmp(&b.score));
     matches
 }