--- a/rust/sorter.rs
+++ b/rust/sorter.rs
@@ -24,36 +24,16 @@ impl Options {
 }
 
 pub fn sort_strings(options: Options, strings: Vec<String>) -> Vec<Match> {
-    let mut matches = Vec::new();
-    let matcher = Arc::new(matcher::Matcher::new(options.pattern));
+    let matcher = matcher::Matcher::new(options.pattern);
 
-    let pool = thread_pool::ThreadPool::new(std::thread::available_parallelism().unwrap().get());
-
-    let (tx, rx) = mpsc::channel::<Match>();
-
-    for string in strings {
-        let thread_matcher = Arc::clone(&matcher);
-        let thread_transmitter = tx.clone();
-        pool.execute(move || {
-            let score = thread_matcher.score(string.to_string());
-            if score > 25 {
-                thread_transmitter
-                    .send(Match {
-                        score,
-                        content: string,
-                    })
-                    .expect("Failed to push data to channel");
-            }
+    let mut matches = strings
+        .into_iter()
+        .map(|candidate| Match {
+            score: matcher.score(candidate.as_str()),
+            content: candidate,
         })
-    }
-
-    drop(pool);
-    drop(tx);
-
-    while let Ok(result) = rx.recv() {
-        matches.push(result)
-    }
-
+        .filter(|m| m.score > 25)
+        .collect::<Vec<Match>>();
     matches.sort_by(|a, b| a.score.cmp(&b.score));
     matches
 }