This PR:

- Adds [Rayon](https://docs.rs/rayon/latest/rayon/) as a dependency
- Adds [Criterion](https://docs.rs/criterion/latest/criterion/) as a dev-dependency
- Adds two Rust/Criterion benchmarks (under `/benches` per convention), directly equivalent to the Lua ones but called without the wrapper
- Adds `rlib` to the generated binary types, to allow for calling `ivyrs` in external crates
- Adds an example usage, mostly for checking output is correct, under `/examples` per convention
- Includes a run of `cargo clippy` and `cargo fmt` over the repo
- Moves the content of the FFI functions out into non-FFI functions, allowing them to be called natively from Rust binaries
- Moves to iterators in place of `for` loops where more idiomatic
- Moves to references instead of owned values where possible
- Restores the use of `minimum_score`
- Removes most use of `std::sync::Mutex`
- Removes the custom thread pool implementation
- Represents an approximate 80% increase in speed in the synthetic benchmarks

See Git commit messages for on-the-way benchmarks

Benchmarks are below, as are before and after flamegraphs of what's going on. It's interesting to note that while the _Rust_ portion of this is significantly faster, the _Lua_ benchmarks don't see much difference. Looks like the bottleneck is not in the library.

# Benchmarks

Everything was run on AMD 5950x, 16c, 32t, 64GB DDR4

## Baseline (Lua)

| Name                           | Total         | Adverage      | Min           | Max           |
|--------------------------------|---------------|---------------|---------------|---------------|
| ivy_match(file.lua) 1000000x   | 01.451465 (s) | 00.000001 (s) | 00.000001 (s) | 00.000018 (s) |
| ivy_files(kubernetes) 100x     | 05.519495 (s) | 00.055195 (s) | 00.041162 (s) | 00.066840 (s) |

## Baseline (Criterion)

| name | label | [Min, Mean, Max] |
|---|---|---|
| ivy_files(kubernetes)  | time:   |[25.538 ms, 25.759 ms, 25.980 ms]|
| ivy_match(file.lua)   |  time:  | [1.1725 µs, 1.1779 µs, 1.1829 µs]|

---

## Final result (Lua)

| Name                           | Total         | Average       | Min           | Max           |
|--------------------------------|---------------|---------------|---------------|---------------|
| ivy_match(file.lua) 1000000x   | 01.384501 (s) | 00.000001 (s) | 00.000001 (s) | 00.000019 (s) |
| ivy_files(kubernetes) 100x     | 05.743775 (s) | 00.057438 (s) | 00.039902 (s) | 00.078781 (s) |

## Final result (Criterion)

| name | label | [Min, Mean, Max] |
|---|---|---|
| ivy_files(kubernetes)  | time:  | [4.5800 ms, 4.6121 ms, 4.6467 ms] |
| ivy_match(file.lua)   |  time:  | [1.1514 µs, 1.1599 µs, 1.1694 µs] |

## Flamegraph prior to changes
![flamegraph-pre-changes](https://user-images.githubusercontent.com/9251567/186944486-88575af1-833f-4feb-9f54-7e9357b74557.svg)

## Flamegraph after changes
![flamegraph-post-changes](https://user-images.githubusercontent.com/9251567/186944587-23309110-4381-46b0-bd4c-1d91ce61592f.svg)